+++ R/renderQCSummary.R
@@ -5,8 +5,6 @@
 #' 
 #' The QC status of all scripts checked into SVN by each author will be displayed, 
 #' along with a high level summary of all scripts that need to be QCed.
-#' 
-#' @param .dirs_exclude Character string (optional). Vector of directories to exclude in the summary (relative to log root).
 #'
 #' @param .output_dir Character string (optional). Path to the directory where the output PDF 
 #'   should be saved. If not provided, the document will not be saved locally.
@@ -23,9 +21,7 @@
 #' }
 #'
 #' @export
-renderQCSummary <- function(.dirs_exclude = NULL, .output_dir = NULL) {
-  
-  dirSummaryRes <- dirSummary(.dirs_exclude = .dirs_exclude)
+renderQCSummary <- function(.output_dir = NULL) {
   
   if (is.null(.output_dir)) {
     .output_dir <- tempdir()
@@ -38,10 +34,17 @@ renderQCSummary <- function(.dirs_exclude = NULL, .output_dir = NULL) {
   output_file <- paste0("qc-summary-", Sys.Date(), ".pdf")
   output_path <- file.path(.output_dir, output_file)
   
+  params_in <- list(
+    project = basename(logRoot()),
+    repoHistory = repoHistory(),
+    qcLog = logRead(),
+    logPending = logPending()
+  )
+  
   rmarkdown::render(
     input = system.file("templates", "QCSummary.Rmd", package = "review"),
     output_file = output_path,
-    params = list(dirSummaryRes = dirSummaryRes),
+    params = params_in,
     envir = new.env(),
     quiet = TRUE
   )