+++ tests/testthat/test-renderQCSummary.R
@@ -2,6 +2,9 @@ if (Sys.getenv("METWORX_VERSION") != "") {
   
   create_test_svn()
   
+  logAccept("script/data-assembly/da-combine-studies.Rmd")
+  logAssign("script/data-assembly/da-functions.R")
+  
   test_that("renderQCSummary works with valid directory", {
     
     expect_message({